Updated Car model to fit the db semantics more. See dsc.
Since we store the encoded image in the database, we can map that column to a regular Java String, because it can store an arbitrarily long sequence of characters. I also modified the method regarding to the base64 encoding, now it takes an array of raw bytes and sets the image string attribute. I also modified those code snippets that references the original base64 method in the model. Hopefully I didn't broke anything beyond repair. :-)